:pencil2::new: commands [show, show-all]
and :fire: traditional commands [save, load]
